preprocess-upsert-2/requirements.txt
azure-functions
azure-storage-blob
openai
pymongo
PyPDF2
reportlab